In Qt6 QWebEnginepage and QWebEngineView can be imported separately before constructing the QApplication
