+++ src/calibre/gui2/tweak_book/reports.py
@@ -29,7 +29,8 @@ from calibre.ebooks.oeb.polish.report import (
     gather_data, CSSEntry, CSSFileMatch, MatchLocation, ClassEntry,
     ClassFileMatch, ClassElement, CSSRule, LinkLocation)
 from calibre.gui2 import error_dialog, question_dialog, choose_save_file, open_url
-from calibre.gui2.webengine import secure_webengine, RestartingWebEngineView
+from calibre.utils.webengine import secure_webengine
+from calibre.gui2.webengine import RestartingWebEngineView
 from calibre.gui2.tweak_book import current_container, tprefs, dictionaries
 from calibre.gui2.tweak_book.widgets import Dialog
 from calibre.gui2.progress_indicator import ProgressIndicator
